Inluce sagas example
