+++ package.json
@@ -67,6 +67,7 @@
     "redux-devtools-extension": "2.13.0",
     "redux-form": "7.1.1",
     "redux-persist": "4.10.2",
+    "redux-saga": "0.16.0",
     "redux-thunk": "2.2.0",
     "sphere-knn": "1.3.1",
     "supercluster": "2.3.0",
